Windows related code reverts
Various changes were reverted to 3.0.16 versions of source given the
faiure to compile on windows. This code primarily is associated with
declaring memory as read only segments. I expect that these code updates
are no longer required once a newer version of NASM is used as an
assembly tool on windows build environments.

Signed-off-by: Jason Katonica <katonica@us.ibm.com>